JOB-101 Remapped owner_id to owner_email
This change allows the jobbergate application to use email as the owner
indicator for its entities. This allows the data migrated from legacy
jobbergate to be immediately assigned to the correct user. With the
custom claims added to the tokens issued by Auth0, jobbergate can handle
all of the filtering and permission control. Further, the data is
explorable without having the overhead of needing to map the Auth0 "sub"
claim to an existing user. See the issue on ClickUp for a through
explanation of the use-case.

* Also updated the slurp project to user owner email
* Also updated cli to use owner email
* Also updated unit tests
* Also updated docker-compose.yaml
* Also renamed "jobbergateapi2" in jobbergate-api to just "jobbergate_api"