JOB-79 Added application_uploaded column to applications
This will be used to indicate if a zip file has been uploaded for the
application. Now that the file is not included in the create endpoint,
it must be separately uploaded.

The upload method will now mark the application to indicate this.

Also updated unit tests.
Also added a delete endpoint and tests for it to remove the application
file.