PENG-1910 Fixed recurring failures in CI pipeline
I moved the session management code in the storage module into its own
context manager. Then, I modified the conftest to override this
particular part of the code only. Now, the storage code is minimally
aware of special test logic (it only blows up if the session context
manager is used in a test) and the particulars of managing a session for
a test are isolated in the conftest.py

I also eliminated the async specific tests. Because the test fixtures
force routes to use a preset session, you can't call multiple routes
simultaneously due to this sqlalchemy error: https://docs.sqlalchemy.org/en/20/errors.html#error-isce.
Any validation of concurrency should occur as part of an integration
test instead.