PENG-2064 Capture job state info in jobbergate
The fundamental feature of this change is to capture the slurm job state
in the API so that it can be observed by users. The major change here is
a remapping of the job_submissions statuses so that they no longer
overlap with job_states reported by slurm and instead are only used to
manage how jobbergate components interact with job submissions along
their lifecycle. New columns are added for the job_info and the
job_state. These capture the details of a job's state as it is
processed.

Change Summary
 - Jobbergate API
   - Added columns for slurm_job_state and slurm_job_info
   - Added mappings for slurm_job_state details (long descriptions and
     abbreviations)
   - Updated schemas for job payloads
   - Created specialized submit and reject endpoints so that agent
     doesn't need to understand job_submission statuses
   - Added a migration for the database changes including existing status mapping
   - Updated and added unit tests
 - Jobbergate Agent
   - Renamed the finish.py module to update.py
   - Changed logic such that all active jobs have their job state
     updated on each pass
   - Moved all logic from the api.py module to update.py and submit.py
   - Updated schemas for revised job payloads
   - Updated and added unit tests
 - Jobbergate CLI changes
   - Updated schemas for revised job payloads
   - Made slurm_job_info hidden by default on job_submissions
 - Jobbergate Docs changes
   - Slight docs updates for the agent app
   - Slight docs updates for the tutorial