PENG-2420 Fix the *snapstore* job in the *Publish on Tag* workflow.
This commit fixes the workflow that publishes the Jobbergate Agent snap. The action *snapcore/action-build* is no longer
used and was taken place by manual build and publish commands. As well as, to keep consistency across the workflows,
the workflow that published the Jobbergate API to ECR is merged in the *Publish on Tag* workflow.