PENG-2426 Fixed file uploads with s3
Jobbergate uses MinIO for file storage internally. The s3 clients use
environment variables for the AWS secrets. These secrets are bound to
the MinIO instance. They will not work for s3 in general. When we were
trying to pull files by URL from s3, our clients were trying to use the
credentials stored in the environment. These credentials were not known
by s3.

Instead of doing all that, we make the assumption that download-by-url
will require the files to be available publicly. Consequently, we can
construct a normal https URL to download the file.

This change applies that logic and fixes the unit tests.