fix(agent): PENG-2457 modify the time param used for querying InfluxDB
This commit modifies the query made against InfluxDB when fetching for
job metrics in a way that the *time* argument is given in nanoseconds.
This is necessary because the agent sends the time in seconds, so, when
fetching the *max_times*, the API will send the time in seconds as well.

However, the data stored in InfluxDB uses the epoch in nanoseconds, which
requires the conversion before executing the query, otherwise all data in
the database will be returned.