cherry-pick: PENG-2678 Fixed issue with handle_errors_async
The implementation of `handle_errors_async()` that was embedded in the
`jobbergate-agent` was not compatible with a recent update to `py-buzz`
version `4.1`.

This change removes the embedded implementation in favor of using the
upstream implementation of the same that is now a part of `py-buzz`
version `5.0`.

Original Commit: 0948b698a3af77865ecea4cf41e9d675967f5cd4