fix(jobbergate-agent-snap): Fixed problem config hook (#696)
When the `snap set` command is used in quick succession (such as when
invoked through the `cloud-init.yaml` of the democluster, the command
can fail. This is happening because the snap attempts to restart rapidly
when it fails to start. The start will fail because the jobbergate-agent
is missing required config.

When the configure hook directs the snap to restart, it may happen while
the snap is in a restart cycle. Snap will detect that the restart
request happened too soon after a failed restart attempt and the command
will fail.

Setting the snap as disabled while the configure hook is running seems
to fix this issue.