chore: add towncrier as changelog manager (#690)
Unreleased entries are now kept on different news fragment files, minimizing conflicts and streamlining the development workflow. Fragments are consolidated into the changelog and deleted when new releases are prepared. There is no major difference when compared to the current way of working, the exact same text we already include under the "Unreleased" header will just go to a different file.