fix(agent): prevent int64 overflow from values fecthed from InfluxDB. (#740)
This commit modifies the `fetch_influx_data` function to return 0 in the measurement
value and the original value is bigger than 2^63 - 1. This prevents an error in the API
where the value is bigger than int64, causing the metrics to not be stored in the database.

(cherry picked from commit a921df564266377dc8eb7e2ada38caf3b8a44f22)